LayoutTests/fast/scrolling/ios/click-events-during-momentum-scroll-in-overflow-after-tap-on-body-expected.txt

Tap here
Verifies that tapping outside of a subscrollable region during momentum scrolling dispatches click events. To run the test manually, swipe up on the black box to scroll it; while scrolling, tap the red box and verify that a click event is dispatched.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS scrolled to minimum threshold
PASS observedClick is true
PASS successfullyParsed is true

TEST COMPLETE
